requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.8.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
docx==0.2.4
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
Flask==3.0.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nltk==3.9.1
numpy==1.26.4
opencv-python==4.11.0.86
orjson==3.11.2
pandas==2.3.0
php==1.2.1
pillow==11.2.1
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
scikit-learn==1.6.1
scipy==1.15.3
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.0.3
